+++ src/common/get_README.ts
@@ -4,20 +4,54 @@
 
 import type { Core } from '../core';
 
-export function get_README(core: Core): string {
+
+type Badges = {
+  prWelcome?: boolean;
+  /** Types included */
+  typescript?: boolean;
+  /** npm package version and downloads/week */
+  npm?: boolean;
+}
+
+export type get_README_Options = {
+  /** All badges default to false. */
+  badges?: Badges
+}
+
+// Be sure when changing the defaults.
+export function get_README(core: Core, options?: get_README_Options): string {
+
+  const projectName = core.consts.projectName;
+
+  function getBadgesString() {
+    const badges = options?.badges;
+    const badgesStrings = [];
+    if (badges?.prWelcome) badgesStrings.push(
+      `[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)`,
+    );
+    if (badges?.typescript) badgesStrings.push(
+      `[![TypeScript](https://badgen.net/npm/types/env-var)](http://www.typescriptlang.org/)`,
+    );
+    if (badges?.npm) badgesStrings.push(...[
+      `[![npm](https://img.shields.io/npm/v/${projectName})](https://www.npmjs.com/package/${projectName})`,
+      `[![npm](https://img.shields.io/npm/dw/${projectName})](https://www.npmjs.com/package/${projectName})`,
+    ]);
+
+    if (!badgesStrings.length)
+      return '';
+      // Must have an empty line between html tags and markdown.
+    return (`<div align="center">\n\n${badgesStrings.join('\n')}\n</div>`);
+  }
+
+
   return (
     `
 <h1 align="center">
-  <!-- <img src=".logo.png" alt=${core.consts.projectName}/><br/> -->
-  ${core.consts.projectName}
+  <!-- <img src=".logo.png" alt=${projectName}/><br/> -->
+  ${projectName}
 </h1>
 
-<div align="center">
-
-  [![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)
-  [![TypeScript](https://badgen.net/npm/types/env-var)](http://www.typescriptlang.org/)
-
-</div>
+${getBadgesString()}
 
 <br/>